+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # PrivateBin version history
 
 ## 2.0.0 (not yet released)
-* CHANGED: Remove page template (#265)
+* CHANGED: Removed page template (#265)
+* CHANGED: Removed support for ZeroBin & v1 pastes - since release 1.3 the v2 format is used (#551)
+* CHANGED: Removed use of base64 & rawinflate libraries (#551)
 * FIXED: Name mismatches in attached files (#1584)
 * FIXED: Unable to paste attachments from clipboard (#1589)